libs/document_storage/pyproject.toml
[tool.pytest.ini_options]
addopts = "--html='../reports/document_storage/unittests/html/index.html' --junitxml='../reports/document_storage/unittests/junit.xml'"

[tool.poetry]
name = "document_storage"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = "Proprietary"
readme = "README.md"

  [[tool.poetry.packages]]
  include = "src"

  [tool.poetry.dependencies]
  python = "^3.12"
  pika = "^1.3.2"
  google-cloud-storage = "^2.15.0"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
